requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
pydantic>=2.7
pandas>=2.2
python-dotenv>=1.0
openai>=1.40
